feat: full toggletip implementation
