+++ src/_data/definitions.json
@@ -2,49 +2,31 @@
   "roles": [
     {
       "term": "Consultants",
-      "definition": "People with lived experience of disability, who are independent or a part of a disability organization, who will consult with Entities on their accessibility plan. This also includes caregivers and parents.",
-      "variations": [
-        "consultant",
-        "individual consultant"
-      ]
+      "definition": "People with lived experience of disability, who are independent or a part of a disability organization, who will consult with Entities on their accessibility plan. This also includes caregivers and parents."
+    },
+    {
+      "term": "Constituency groups",
+      "definition": "TODO."
     },
     {
       "term": "Representative consultants",
-      "definition": "Consultants who represent a constituency group. They are appointed by the group, or matched by the Social Enterprise.",
-      "variations": [
-        "representatives",
-        "representative",
-        "representative consultant"
-      ]
+      "definition": "Consultants who represent a constituency group. They are appointed by the group, or matched by the Social Enterprise."
     },
     {
       "term": "Federally regulated entities",
-      "definition": "Any organization that is regulated by the federal government under the Accessible Canada Act.",
-      "variations": [
-        "entity",
-        "entities"
-      ]
+      "definition": "Any organization that is regulated by the federal government under the Accessible Canada Act."
     },
     {
       "term": "Disability organizations",
-      "definition": "Any organization that serves the disability community. ",
-      "variations": [
-        "disability organization"
-      ]
+      "definition": "Any organization that serves the disability community. "
     },
     {
       "term": "Caregivers",
-      "definition": "This can be a family member, friend, or someone trained to provide care. They act as a liaison between entity and consultant, and ensures safety and well-being of the consultant.",
-      "variations": [
-        "caregiver"
-      ]
+      "definition": "This can be a family member, friend, or someone trained to provide care. They act as a liaison between entity and consultant, and ensures safety and well-being of the consultant."
     },
     {
       "term": "Support people",
-      "definition": "This can be anyone the consultant has picked as someone who can provide support to them. The support person does not act as the liaison.",
-      "variations": [
-        "support person"
-      ]
+      "definition": "This can be anyone the consultant has picked as someone who can provide support to them. The support person does not act as the liaison."
     },
     {
       "term": "Social enterprise",
@@ -53,24 +35,15 @@
     },
     {
       "term": "Outreach team",
-      "definition": "People who are members of the disability community and/or other under-represented communities, who will do outreach to potential consultants and support them to become consultants",
-      "variations": [
-        "outreach team member"
-      ]
+      "definition": "People who are members of the disability community and/or other under-represented communities, who will do outreach to potential consultants and support them to become consultants"
     },
     {
       "term": "Advisory committee",
-      "definition": "A trained group of people with lived experience of disability and intersectionality. They ensure representation of consultants on projects, and provides and mentorship to the consultants during the consultation process.",
-      "variations": [
-        "advisory committee member"
-      ]
+      "definition": "A trained group of people with lived experience of disability and intersectionality. They ensure representation of consultants on projects, and provides and mentorship to the consultants during the consultation process."
     },
     {
       "term": "Third-party conveners",
-      "definition": "Trained individuals who conducts and facilitates the consultation the behalf of Federally Regulated Entities.",
-      "variations": [
-        "convener"
-      ]
+      "definition": "Trained individuals who conducts and facilitates the consultation the behalf of Federally Regulated Entities."
     }
   ]
 }
