+++ src/utils/getEntry.js
@@ -4,11 +4,11 @@ const definitions = require("../_data/definitions.json");
 
 module.exports = term => {
     const matches = definitions.roles.filter(entry => {
-        return term.toLowerCase() === entry.term.toLowerCase() || entry.variations.includes(term.toLowerCase());
+        return term.toLowerCase() === entry.term.toLowerCase();
     });
 
     if (matches.length > 0) {
-        return matches[0].definition;
+        return matches[0];
     }
 
     return false;
