+++ src/utils/getTerms.js
@@ -5,6 +5,6 @@ const definitions = require("../_data/definitions.json");
 module.exports = () => {
     return definitions.roles.reduce(function (accumulator, currentValue) {
         accumulator.push(currentValue.term);
-        return accumulator.concat(currentValue.variations);
+        return accumulator;
     }, []);
 };
